Projects/05-Voice_Assistant/requirements.txt
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.0
elevenlabs>=0.2.0 